CONN_2161 Modified HBM files for new schema, refactored unit tests and classes to match changes. Implemented DAOIntegrationTest base for JUnit
