+++ Product/Production/Adapters/DocumentSubmission_a0/src/main/java/gov/hhs/fha/nhinc/docsubmission/adapter/component/AdapterComponentDocSubmissionImpl.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2009-2018, United States Government, as represented by the Secretary of Health and Human Services.
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *     * Redistributions of source code must retain the above
@@ -12,7 +12,7 @@
  *     * Neither the name of the United States Government nor the
  *       names of its contributors may be used to endorse or promote products
  *       derived from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -40,17 +40,17 @@ import oasis.names.tc.ebxml_regrep.xsd.rs._3.RegistryResponseType;
 public class AdapterComponentDocSubmissionImpl extends BaseService {
 
     public RegistryResponseType provideAndRegisterDocumentSetb(AdapterProvideAndRegisterDocumentSetRequestType body,
-            WebServiceContext context) {
+        WebServiceContext context) {
         AssertionType assertion = getAssertion(context, body.getAssertion());
 
-        return new AdapterComponentDocSubmissionOrchImpl().provideAndRegisterDocumentSetB(
-                body.getProvideAndRegisterDocumentSetRequest(), assertion);
+        return AdapterComponentDocSubmissionOrchImpl.provideAndRegisterDocumentSetB(
+            body.getProvideAndRegisterDocumentSetRequest(), assertion);
     }
 
     public RegistryResponseType provideAndRegisterDocumentSetb(ProvideAndRegisterDocumentSetRequestType body,
-            WebServiceContext context) {
+        WebServiceContext context) {
         AssertionType assertion = getAssertion(context, null);
 
-        return new AdapterComponentDocSubmissionOrchImpl().provideAndRegisterDocumentSetB(body, assertion);
+        return AdapterComponentDocSubmissionOrchImpl.provideAndRegisterDocumentSetB(body, assertion);
     }
 }
