+++ Product/Production/Adapters/General/CONNECTAdminGUI/src/main/java/gov/hhs/fha/nhinc/admingui/services/LoadTestDataService.java
@@ -27,7 +27,7 @@
 package gov.hhs.fha.nhinc.admingui.services;
 
 import gov.hhs.fha.nhinc.admingui.services.exception.LoadTestDataException;
-import gov.hhs.fha.nhinc.docrepository.adapter.model.Document;
+import gov.hhs.fha.nhinc.docrepository.adapter.model.DocumentMetadata;
 import gov.hhs.fha.nhinc.docrepository.adapter.model.EventCode;
 import gov.hhs.fha.nhinc.patientdb.model.Address;
 import gov.hhs.fha.nhinc.patientdb.model.Identifier;
@@ -84,21 +84,21 @@ public interface LoadTestDataService {
 
     public boolean savePhonenumber(Phonenumber phonenumber) throws LoadTestDataException;
 
-    public List<Document> getAllDocuments();
+    public List<DocumentMetadata> getAllDocuments();
 
     public List<EventCode> getAllEventCodesBy(Long documentId);
 
-    public Document duplicateDocument(Long documentId);
+    public DocumentMetadata duplicateDocument(Long documentId);
 
-    public boolean deleteDocument(Document document);
+    public boolean deleteDocument(DocumentMetadata document);
 
     public boolean deleteEventCode(EventCode eventCode);
 
-    public Document getDocumentBy(Long documentId);
+    public DocumentMetadata getDocumentBy(Long documentId);
 
     public EventCode getEventCodeBy(Long eventCodeId);
 
-    public boolean saveDocument(Document document) throws LoadTestDataException;
+    public boolean saveDocument(DocumentMetadata document) throws LoadTestDataException;
 
     public boolean saveEventCode(EventCode eventCode) throws LoadTestDataException;
