+++ Product/Production/Common/CONNECTCoreLib/pom.xml
@@ -307,6 +307,13 @@
             <artifactId>encoder</artifactId>
             <version>1.2.1</version>
     </dependency>
+	<dependency>
+	    <groupId>mysql</groupId>
+	    <artifactId>mysql-connector-java</artifactId>
+	    <version>5.1.10</version>
+	    <scope>test</scope>
+	</dependency>
+    
         
     </dependencies>
