+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/docrepository/adapter/dao/DocumentDao.java
@@ -26,18 +26,19 @@
  */
 package gov.hhs.fha.nhinc.docrepository.adapter.dao;
 
-import gov.hhs.fha.nhinc.docrepository.adapter.model.Document;
+import gov.hhs.fha.nhinc.docrepository.adapter.model.DocumentMetadata;
 import gov.hhs.fha.nhinc.docrepository.adapter.model.DocumentQueryParams;
 import gov.hhs.fha.nhinc.persistence.HibernateUtilFactory;
 import gov.hhs.fha.nhinc.util.GenericDBUtils;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import org.apache.commons.collections.CollectionUtils;
 import org.hibernate.Criteria;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.criterion.Criterion;
-import org.hibernate.criterion.Expression;
 import org.hibernate.criterion.Restrictions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -50,40 +51,41 @@ import org.slf4j.LoggerFactory;
 public class DocumentDao {
     private static final Logger LOG = LoggerFactory.getLogger(DocumentDao.class);
 
-    public boolean save(Document document) {
+    public boolean save(DocumentMetadata document) {
         return GenericDBUtils.save(getSession(), document);
     }
 
-    public Document saveAndGetDocument(Document document) {
-        Document s = null;
+    public DocumentMetadata saveAndGetDocument(DocumentMetadata document) {
+        DocumentMetadata s = null;
         if (GenericDBUtils.save(getSession(), document)) {
             s = findById(document.getDocumentid());
         }
         return s;
     }
 
-    public boolean saveAll(List<Document> documents) {
+    public boolean saveAll(List<DocumentMetadata> documents) {
         return GenericDBUtils.saveAll(getSession(), documents);
     }
 
-    public boolean delete(Document document) {
+    public boolean delete(DocumentMetadata document) {
         return GenericDBUtils.delete(getSession(), document);
     }
 
-    public boolean deleteAll(List<Document> documents) {
+    public boolean deleteAll(List<DocumentMetadata> documents) {
         return GenericDBUtils.deleteAll(getSession(), documents);
     }
 
-    public Document findById(Long documentId) {
-        return GenericDBUtils.readBy(getSession(), Document.class, documentId);
+    public DocumentMetadata findById(Long documentId) {
+        return GenericDBUtils.readBy(getSession(), DocumentMetadata.class, documentId);
     }
 
-    public List<Document> findAll() {
-        return GenericDBUtils.findAll(getSession(), Document.class);
+    public List<DocumentMetadata> findAll() {
+        return GenericDBUtils.findAll(getSession(), DocumentMetadata.class);
     }
 
-    public List<Document> findAllBy(long patientId) {
-        return GenericDBUtils.findAllBy(getSession(), Document.class, Expression.eq("patientRecordId", patientId));
+    public List<DocumentMetadata> findAllBy(long patientId) {
+        return GenericDBUtils.findAllBy(getSession(), DocumentMetadata.class,
+            Restrictions.eq("patientRecordId", patientId));
     }
 
     protected Session getSession() {
@@ -102,8 +104,7 @@ public class DocumentDao {
      * @param params Query parameters
      * @return Query results
      */
-    @SuppressWarnings("unchecked")
-    public List<Document> findDocuments(DocumentQueryParams params) {
+    public List<DocumentMetadata> findDocuments(DocumentQueryParams params) {
         LOG.debug("Beginning document query");
 
         String patientId = null;
@@ -117,6 +118,7 @@ public class DocumentDao {
         Date serviceStopTimeTo = null;
         List<String> statuses = null;
         List<String> documentUniqueIds = null;
+        Boolean onDemand = null;
         if (params != null) {
             patientId = params.getPatientId();
             classCodes = params.getClassCodes();
@@ -129,149 +131,133 @@ public class DocumentDao {
             serviceStopTimeTo = params.getServiceStopTimeTo();
             statuses = params.getStatuses();
             documentUniqueIds = params.getDocumentUniqueIds();
+            onDemand = params.getOnDemand();
         }
-        List<Document> documents = null;
+        List<DocumentMetadata> documents = new ArrayList<>();
         Session sess = null;
         try {
             sess = getSession();
             if (sess != null) {
                 SimpleDateFormat logDateFormatter = new SimpleDateFormat("yyyyMMdd hh:mm:ss a");
-                Criteria criteria = sess.createCriteria(Document.class);
+                Criteria criteria = sess.createCriteria(DocumentMetadata.class);
 
                 if (patientId != null) {
-                    criteria.add(Expression.eq("patientId", patientId));
+                    criteria.add(Restrictions.eq("patientId", patientId));
                 }
 
-                if (classCodes != null && !classCodes.isEmpty()) {
-                    /**************************************************************
-                     * The class code and class code scheme combination can come in two different formats:
-                     *
-                     * <ns7:Slot name="$XDSDocumentEntryClassCode"> <ns7:ValueList> <ns7:Value>34133-9</ns7:Value>
-                     * </ns7:ValueList> </ns7:Slot> <ns7:Slot name="$XDSDocumentEntryClassCodeScheme"> <ns7:ValueList>
-                     * <ns7:Value>2.16.840.1.113883.6 .1</ns7:Value> </ns7:ValueList> </ns7:Slot>
-                     *
-                     * or
-                     *
-                     * <ns7:Slot name="$XDSDocumentEntryClassCode"> <ns7:ValueList> <ns7:Value>(
-                     * '34133-9^^2.16.840.1.113883.6.1')</ns7:Value> </ns7:ValueList> </ns7:Slot>
-                     *
-                     * The code below can deal with both formats.
-                     *
-                     *************************************************************/
-                    Criterion criterion = null;
-                    for (String classCode : classCodes) {
-                        if (LOG.isDebugEnabled()) {
-                            LOG.debug("Document query - class code: " + classCode);
-                        }
-                        String newClassCode;
-                        String newCodeScheme;
-
-                        if (classCode.contains("^^")) {
-                            int index = classCode.indexOf("^^");
-                            newClassCode = classCode.substring(0, index);
-                            newCodeScheme = classCode.substring(index + 2);
-                        } else {
-                            newClassCode = classCode;
-                            newCodeScheme = classCodeScheme;
-                        }
-
-                        Criterion andCrit = Expression.eq("classCode", newClassCode);
-                        if (newCodeScheme != null && !newCodeScheme.isEmpty()) {
-                            andCrit = Restrictions.and(andCrit, Expression.eq("classCodeScheme", newCodeScheme));
-                        }
-                        if (criterion == null) {
-                            criterion = andCrit;
-                        } else {
-                            criterion = Restrictions.or(criterion, andCrit);
-                        }
-                    }
-                    criteria.add(criterion);
-                }
+                getClassCodeCritera(classCodes, classCodeScheme, criteria);
 
                 if (creationTimeFrom != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug("Document query - creation time from: " + logDateFormatter.format(creationTimeFrom));
-                    }
-                    criteria.add(Expression.ge("creationTime", creationTimeFrom));
+                    LOG.debug("Document query - creation time from: {}", logDateFormatter.format(creationTimeFrom));
+                    criteria.add(Restrictions.ge("creationTime", creationTimeFrom));
                 }
 
                 if (creationTimeTo != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug("Document query - creation time to: " + logDateFormatter.format(creationTimeTo));
-                    }
-                    criteria.add(Expression.le("creationTime", creationTimeTo));
+                    LOG.debug("Document query - creation time to: {}", logDateFormatter.format(creationTimeTo));
+                    criteria.add(Restrictions.le("creationTime", creationTimeTo));
                 }
 
                 if (serviceStartTimeFrom != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug("Document query - service start time from: "
-                            + logDateFormatter.format(serviceStartTimeFrom));
-                    }
-                    criteria.add(Expression.ge("serviceStartTime", serviceStartTimeFrom));
+                    LOG.debug("Document query - service start time from: {}",
+                        logDateFormatter.format(serviceStartTimeFrom));
+                    criteria.add(Restrictions.ge("serviceStartTime", serviceStartTimeFrom));
                 }
 
                 if (serviceStartTimeTo != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug(
-                            "Document query - service start time to: " + logDateFormatter.format(serviceStartTimeTo));
-                    }
-                    criteria.add(Expression.le("serviceStartTime", serviceStartTimeTo));
+                    LOG.debug("Document query - service start time to: {}",
+                        logDateFormatter.format(serviceStartTimeTo));
+                    criteria.add(Restrictions.le("serviceStartTime", serviceStartTimeTo));
                 }
 
                 if (serviceStopTimeFrom != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug(
-                            "Document query - service stop time from: " + logDateFormatter.format(serviceStopTimeFrom));
-                    }
-                    criteria.add(Expression.ge("serviceStopTime", serviceStopTimeFrom));
+                    LOG.debug("Document query - service stop time from: {}",
+                        logDateFormatter.format(serviceStopTimeFrom));
+                    criteria.add(Restrictions.ge("serviceStopTime", serviceStopTimeFrom));
                 }
 
                 if (serviceStopTimeTo != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug(
-                            "Document query - service stop time to: " + logDateFormatter.format(serviceStopTimeTo));
-                    }
-                    criteria.add(Expression.le("serviceStopTime", serviceStopTimeTo));
+                    LOG.debug("Document query - service stop time to: {}", logDateFormatter.format(serviceStopTimeTo));
+                    criteria.add(Restrictions.le("serviceStopTime", serviceStopTimeTo));
                 }
 
                 if (statuses != null && !statuses.isEmpty()) {
-                    if (LOG.isDebugEnabled()) {
-                        for (String status : statuses) {
-                            LOG.debug("Document query - status: " + status);
-                        }
-                    }
-                    criteria.add(Expression.in("status", statuses));
+                    LOG.debug("Document query - statuses: {}", statuses);
+                    criteria.add(Restrictions.in("status", statuses));
                 }
 
                 if (documentUniqueIds != null && !documentUniqueIds.isEmpty()) {
-                    if (LOG.isDebugEnabled()) {
-                        for (String documentUniqueId : documentUniqueIds) {
-                            LOG.debug("Document query - document unique id: " + documentUniqueId);
-                        }
-                    }
-                    criteria.add(Expression.in("documentUniqueId", documentUniqueIds));
+                    LOG.debug("Document query - document unique ids: {}", documentUniqueIds);
+                    criteria.add(Restrictions.in("documentUniqueId", documentUniqueIds));
                 }
 
-                if (params.getOnDemand() != null) {
-                    if (LOG.isDebugEnabled()) {
-                        LOG.debug("Document query - onDemand: " + params.getOnDemand());
-                    }
-                    criteria.add(Expression.eq("onDemand", params.getOnDemand()));
+                if (onDemand != null) {
+                    LOG.debug("Document query - onDemand: {}", onDemand);
+                    criteria.add(Restrictions.eq("onDemand", onDemand));
                 }
 
                 documents = criteria.list();
+
+                LOG.debug("Completed retrieve of document query. {} results returned.", documents.size());
             } else {
                 LOG.error("Failed to obtain a session from the sessionFactory");
             }
 
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("Completed retrieve of document query. "
-                    + (documents == null ? "0" : Integer.toString(documents.size())) + " results returned.");
-            }
+
         } finally {
             GenericDBUtils.closeSession(sess);
         }
         return documents;
+    }
+
+    /**************************************************************
+     * The class code and class code scheme combination can come in two different formats:
+     *
+     * <ns7:Slot name="$XDSDocumentEntryClassCode"> <ns7:ValueList> <ns7:Value>34133-9</ns7:Value> </ns7:ValueList>
+     * </ns7:Slot> <ns7:Slot name="$XDSDocumentEntryClassCodeScheme"> <ns7:ValueList> <ns7:Value>2.16.840.1.113883.6
+     * .1</ns7:Value> </ns7:ValueList> </ns7:Slot>
+     *
+     * or
+     *
+     * <ns7:Slot name="$XDSDocumentEntryClassCode"> <ns7:ValueList> <ns7:Value>(
+     * '34133-9^^2.16.840.1.113883.6.1')</ns7:Value> </ns7:ValueList> </ns7:Slot>
+     *
+     * This method deals with both formats.
+     *
+     * Any critera generated with this method will be added to the Criteria parameter
+     *
+     *************************************************************/
+    private static void getClassCodeCritera(List<String> classCodes, String classCodeScheme, Criteria criteria) {
+        if (CollectionUtils.isNotEmpty(classCodes)) {
+            Criterion criterion = null;
+            for (String classCode : classCodes) {
+
+                LOG.debug("Document query - class code: {}", classCode);
+
+                String newClassCode;
+                String newCodeScheme;
+
+                if (classCode.contains("^^")) {
+                    int index = classCode.indexOf("^^");
+                    newClassCode = classCode.substring(0, index);
+                    newCodeScheme = classCode.substring(index + 2);
+                } else {
+                    newClassCode = classCode;
+                    newCodeScheme = classCodeScheme;
+                }
+
+                Criterion andCrit = Restrictions.eq("classCode", newClassCode);
+                if (newCodeScheme != null && !newCodeScheme.isEmpty()) {
+                    andCrit = Restrictions.and(andCrit, Restrictions.eq("classCodeScheme", newCodeScheme));
+                }
+
+                if (criterion == null) {
+                    criterion = andCrit;
+                } else {
+                    criterion = Restrictions.or(criterion, andCrit);
+                }
+            }
+            criteria.add(criterion);
+        }
     }
 
 }
