+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/properties/PropertyAccessor.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2009-2018, United States Government, as represented by the Secretary of Health and Human Services.
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *     * Redistributions of source code must retain the above
@@ -12,7 +12,7 @@
  *     * Neither the name of the United States Government nor the
  *       names of its contributors may be used to endorse or promote products
  *       derived from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -42,6 +42,7 @@ public class PropertyAccessor implements IPropertyAcessor {
 
     private PropertyFileDAO propertyFileDAO;
     private PropertyAccessorFileUtilities fileUtilities;
+    public static final PropertyAccessor INSTANCE = new PropertyAccessor();
 
     private static final Logger LOG = LoggerFactory.getLogger(PropertyAccessor.class);
 
@@ -53,14 +54,9 @@ public class PropertyAccessor implements IPropertyAcessor {
         fileUtilities = createPropertyAccessorFileUtilities();
     }
 
-    private static class SingletonHolder {
-
-        public static final PropertyAccessor INSTANCE = new PropertyAccessor();
-    }
-
     // singleton
     public static PropertyAccessor getInstance() {
-        return SingletonHolder.INSTANCE;
+        return INSTANCE;
     }
 
     /**
