+++ Product/Production/Common/CONNECTCoreLib/src/test/java/gov/hhs/fha/nhinc/docrepository/adapter/dao/DocumentDaoTest.java
@@ -31,7 +31,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import gov.hhs.fha.nhinc.docrepository.adapter.model.Document;
+import gov.hhs.fha.nhinc.docrepository.adapter.model.DocumentMetadata;
 import gov.hhs.fha.nhinc.docrepository.adapter.model.DocumentQueryParams;
 import java.util.ArrayList;
 import java.util.Date;
@@ -62,7 +62,7 @@ public class DocumentDaoTest {
 
     @Test
     public void testSave() {
-        Document doc = new Document();
+        DocumentMetadata doc = new DocumentMetadata();
 
         documentDao.save(doc);
 
@@ -71,7 +71,7 @@ public class DocumentDaoTest {
 
     @Test
     public void testDelete() {
-        Document doc = new Document();
+        DocumentMetadata doc = new DocumentMetadata();
 
         documentDao.delete(doc);
 
@@ -80,12 +80,12 @@ public class DocumentDaoTest {
 
     @Test
     public void testDeleteAll() {
-        List<Document> documents = new ArrayList<>();
-        Document document1 = new Document();
+        List<DocumentMetadata> documents = new ArrayList<>();
+        DocumentMetadata document1 = new DocumentMetadata();
         documents.add(document1);
-        Document document2 = new Document();
+        DocumentMetadata document2 = new DocumentMetadata();
         documents.add(document2);
-        Document document3 = new Document();
+        DocumentMetadata document3 = new DocumentMetadata();
         documents.add(document3);
 
         when(session.isOpen()).thenReturn(true);
@@ -101,37 +101,37 @@ public class DocumentDaoTest {
 
     @Test
     public void testFindById() {
-        Document doc = new Document();
+        DocumentMetadata doc = new DocumentMetadata();
         final long ID = 12345;
         doc.setDocumentid(ID);
 
-        when(session.get(Document.class, ID)).thenReturn(doc);
+        when(session.get(DocumentMetadata.class, ID)).thenReturn(doc);
 
-        Document resultDoc = documentDao.findById(ID);
+        DocumentMetadata resultDoc = documentDao.findById(ID);
 
         assertEquals(resultDoc.getDocumentid(), (Long) ID);
     }
 
     @Test
     public void testFindAll() {
-        Document doc1 = new Document();
+        DocumentMetadata doc1 = new DocumentMetadata();
         final long ID_1 = 12345;
         doc1.setDocumentid(ID_1);
 
-        Document doc2 = new Document();
+        DocumentMetadata doc2 = new DocumentMetadata();
         final long ID_2 = 12345;
         doc2.setDocumentid(ID_2);
 
-        List<Document> documents = new ArrayList<>();
+        List<DocumentMetadata> documents = new ArrayList<>();
         documents.add(doc1);
         documents.add(doc2);
 
         Criteria criteria = mock(Criteria.class);
 
-        when(session.createCriteria(Document.class)).thenReturn(criteria);
+        when(session.createCriteria(DocumentMetadata.class)).thenReturn(criteria);
         when(criteria.list()).thenReturn(documents);
 
-        List<Document> resultsDocuments = documentDao.findAll();
+        List<DocumentMetadata> resultsDocuments = documentDao.findAll();
 
         assertEquals(resultsDocuments.size(), 2);
         assertEquals(resultsDocuments.get(0).getDocumentid(), (Long) ID_1);
@@ -173,27 +173,27 @@ public class DocumentDaoTest {
         params.setDocumentUniqueId(documentUniqueIds);
         params.setOnDemandParams(false);
 
-        Document doc1 = new Document();
+        DocumentMetadata doc1 = new DocumentMetadata();
         final long ID_1 = 12345;
         doc1.setDocumentid(ID_1);
 
-        Document doc2 = new Document();
+        DocumentMetadata doc2 = new DocumentMetadata();
         final long ID_2 = 12345;
         doc2.setDocumentid(ID_2);
 
-        List<Document> documents = new ArrayList<>();
+        List<DocumentMetadata> documents = new ArrayList<>();
         documents.add(doc1);
         documents.add(doc2);
 
         Criteria criteria = mock(Criteria.class);
 
-        when(session.createCriteria(Document.class)).thenReturn(criteria);
+        when(session.createCriteria(DocumentMetadata.class)).thenReturn(criteria);
         when(criteria.list()).thenReturn(documents);
 
-        when(session.createCriteria(Document.class)).thenReturn(criteria);
+        when(session.createCriteria(DocumentMetadata.class)).thenReturn(criteria);
         when(criteria.list()).thenReturn(documents);
 
-        List<Document> resultsDocuments = documentDao.findDocuments(params);
+        List<DocumentMetadata> resultsDocuments = documentDao.findDocuments(params);
 
         assertEquals(resultsDocuments.size(), 2);
         assertEquals(resultsDocuments.get(0).getDocumentid(), (Long) ID_1);
