+++ Product/Production/Common/CONNECTCoreLib/src/test/resources/config/CONNECT-testcontext.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:aop="http://www.springframework.org/schema/aop"
+       xmlns:tx="http://www.springframework.org/schema/tx"
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:jee="http://www.springframework.org/schema/jee"
+       xsi:schemaLocation="
+        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd
+        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+        http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.0.xsd
+        http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
+        http://www.springframework.org/schema/jee http://www.springframework.org/schema/jee/spring-jee-3.0.xsd">
+
+    <bean name="eventfactory" class="gov.hhs.fha.nhinc.event.EventFactoryFactoryBean" />
+
+    <bean name="txHibernateUtil" class="gov.hhs.fha.nhinc.logging.transaction.persistance.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+    <bean name="eventHibernateUtil" class="gov.hhs.fha.nhinc.event.persistence.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+    <bean name="asyncmsgsHibernateUtil" class="gov.hhs.fha.nhinc.asyncmsgs.persistence.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+    <bean name="connManHibernateUtil" class="gov.hhs.fha.nhinc.common.connectionmanager.persistence.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+    <bean name="docRepoHibernateUtil" class="gov.hhs.fha.nhinc.docrepository.adapter.persistence.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+    <bean name="patientCorrHibernateUtil" class="gov.hhs.fha.nhinc.patientcorrelation.nhinc.persistence.HibernateUtil"
+        init-method="buildSessionFactory" destroy-method="closeSessionFactory" lazy-init="false"/>
+
+
+</beans>
