+++ Product/Production/Common/CONNECTCoreLib/src/test/resources/config/hibernate/dyndocrepo.hibernate.cfg.xml
@@ -1,0 +1,26 @@
+<?xml version='1.0' encoding='utf-8'?>
+<!DOCTYPE hibernate-configuration SYSTEM "classpath://org/hibernate/hibernate-configuration-3.0.dtd">
+
+<hibernate-configuration>
+
+    <session-factory>
+
+        <!-- SQL dialect -->
+        <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
+
+        <!-- Disable the second-level cache  -->
+        <property name="hibernate.cache.provider_class">org.hibernate.cache.internal.NoCachingRegionFactory</property>
+
+        <property name="hibernate.hbm2ddl.auto">update</property>
+
+        <property name="hibernate.connection.datasource">jdbc/docrepository_datasource</property>
+		
+		
+
+        <mapping resource="Document.hbm.xml"/>
+        <mapping resource="EventCode.hbm.xml"/>
+        <mapping resource="ExtraSlot.hbm.xml"/>
+
+    </session-factory>
+
+</hibernate-configuration>
