+++ Product/Production/Common/CONNECTCoreLib/src/test/resources/config/hibernate/event.hibernate.cfg.xml
@@ -1,0 +1,23 @@
+<?xml version='1.0' encoding='utf-8'?>
+<!DOCTYPE hibernate-configuration SYSTEM "classpath://org/hibernate/hibernate-configuration-3.0.dtd">
+
+<hibernate-configuration>
+
+    <session-factory>
+		<property name="hibernate.show_sql">true</property>
+        <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
+
+        <!-- Uncomment the following line if you want the HBM files to update the database tables.
+             Useful for when you are chaning the schema and want to make sure everything works -->
+        <!-- property name="hibernate.hbm2ddl.auto">update</property -->
+
+        <property name="hibernate.current_session_context_class">thread</property>
+        <property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
+        <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/eventdb</property>
+        <property name="hibernate.connection.username">root</property>
+        <property name="hibernate.connection.password">NHIE-Gateway</property>
+		
+        <mapping resource="event.hbm.xml"/>
+    </session-factory>
+
+</hibernate-configuration>
