+++ Product/Production/Services/DocumentSubmissionCore/src/main/java/gov/hhs/fha/nhinc/docsubmission/adapter/component/XDRHelper.java
@@ -219,7 +219,7 @@ public class XDRHelper {
                     SlotType1 recipSlot = getNamedSlotItem(extObj.getSlot(), XDS_INTENDED_RECIPIENT_SLOT);
                     if (recipSlot != null) {
                         result = recipSlot.getValueList().getValue();
-                        LOG.info(result.toString());
+                        LOG.info("Intended Recipients: {}", result.toString());
                     }
 
                 }
